Split benches test into another file
